chore: refine debug endpoint
